feat: implement custom dialog for delete in daily works and time logs

+++ src/app/routes/daily-works/daily-works.component.ts
@@ -8,11 +8,13 @@ import {
   signal,
 } from '@angular/core';
 import { rxResource } from '@angular/core/rxjs-interop';
-import { filter } from 'rxjs';
+import { filter, firstValueFrom, map } from 'rxjs';
 import { PagerComponent } from '../../shared/components/pager/pager.component';
 import { QuickInsertDialogComponent } from '../../shared/dialogs/quick-insert-dialog/quick-insert-dialog.component';
 import { DailyWorksService } from '../../shared/services/daily-works.service';
 import { TimeLogsService } from '../../shared/services/time-logs.service';
+import { CustomDialogComponent } from '../../shared/dialogs/custom-dialog/custom-dialog.component';
+import { CustomDialogService } from '../../shared/services/custom-dialog.service';
 
 @Component({
   imports: [DatePipe, DialogModule, PagerComponent],
@@ -24,6 +26,7 @@ export class DailyWorksComponent {
   private readonly dailyWorksService = inject(DailyWorksService);
   private readonly timeLogsService = inject(TimeLogsService);
   private readonly dialog = inject(Dialog);
+  private readonly customDialogService = inject(CustomDialogService);
 
   readonly currentPage = signal<number>(1);
   readonly currentPageSize = signal<number>(10);
@@ -69,13 +72,23 @@ export class DailyWorksComponent {
     });
   }
 
-  deleteRow(day: string) {
-    this.timeLogsService.deleteTimeLogsByDate(day).subscribe({
-      next: () => {
-        this.totalRowsResource.reload();
-        this.dailyWorksResource.reload();
-      },
-      error: (error) => console.error('Errore:', error),
-    });
+  async deleteRow(day: string) {
+    if (
+      await this.customDialogService.show({
+        title: 'Alert',
+        message: 'Are you sure you want to delete this row?',
+        showCancelButton: true,
+        confirmButtonType: 'error',
+        confirmButtonText: 'Delete',
+        cancelButtonText: 'Cancel',
+      })
+    )
+      this.timeLogsService.deleteTimeLogsByDate(day).subscribe({
+        next: () => {
+          this.totalRowsResource.reload();
+          this.dailyWorksResource.reload();
+        },
+        error: (error) => console.error('Errore:', error),
+      });
   }
 }
